feat: move utility function to util file
