+++ js/src/index.ts
@@ -1,0 +1,3 @@
+export const HBP_VERSION = 0x01;
+
+export * from "./Deserialzer.js";
